+++ deploy/helm/mcp-server/values.yaml
@@ -22,6 +22,8 @@ env:
   MCP_TRANSPORT_PROTOCOL: "http"
   PYTHON_LOG_LEVEL: "INFO"
   PROMETHEUS_URL: "https://thanos-querier.openshift-monitoring.svc.cluster.local:9091"
+  TEMPO_URL: "https://tempo-tempostack-gateway.observability-hub.svc.cluster.local:8080"
+  TEMPO_TENANT_ID: "dev"
   MAX_TIME_RANGE_DAYS: 90
   DEFAULT_TIME_RANGE_DAYS: 7
