+++ deploy/helm/observability/tempo/values.yaml
@@ -33,7 +33,7 @@ tempoStack:
       enabled: true
     queryFrontend:
       jaegerQuery:
-        enabled: false
+        enabled: true
         # ingress:
         #   type: route
