+++ src/mcp_server/env.template
@@ -1,6 +1,7 @@
 # Observability MCP Server Configuration
 
 PROMETHEUS_URL=http://localhost:9090
+TEMPO_URL=http://localhost:8082
 THANOS_TOKEN=
 VERIFY_SSL=true
 LOG_LEVEL=INFO
