+++ src/mcp_server/observability_mcp.py
@@ -45,7 +45,7 @@ class ObservabilityMCPServer:
             find_best_metric_with_metadata_v2,  # Smart metric selection v2
             find_best_metric_with_metadata,   # Smart metric selection v1
         )
-        from .tools.tempo import (
+        from .tools.tempo_query_tool import (
             query_tempo_tool,
             get_trace_details_tool,
             chat_tempo_tool,
@@ -67,11 +67,11 @@ class ObservabilityMCPServer:
         self.mcp.tool()(list_openshift_metric_groups)
         self.mcp.tool()(list_openshift_namespace_metric_groups)
         self.mcp.tool()(chat_openshift)
-        
+
         # Register Prometheus tools one by one
         self.mcp.tool()(search_metrics)                    # Search metrics by pattern
         self.mcp.tool()(get_metric_metadata)              # Get metric metadata
-        self.mcp.tool()(get_label_values)                 # Get label values  
+        self.mcp.tool()(get_label_values)                 # Get label values
         self.mcp.tool()(execute_promql)                   # Execute PromQL queries
         self.mcp.tool()(explain_results)                  # Explain query results
         self.mcp.tool()(suggest_queries)                  # Suggest related queries
