--- src/mcp_server/tools/tempo/__init__.py
@@ -1,25 +1,0 @@
-"""Tempo package for trace analysis and querying.
-
-This package provides modular functionality for interacting with Tempo traces:
-- Base Tempo client functionality
-- Query execution and processing
-- Data formatting and display
-- Analysis and pattern detection
-- MCP tool functions
-"""
-
-from .tempo_query_tool import query_tempo_tool, get_trace_details_tool, chat_tempo_tool
-from .tempo_base import TempoQueryTool
-from .tempo_query import TempoQueryService
-from .tempo_formatters import TempoTraceFormatter
-from .tempo_analyzers import TempoTraceAnalyzer
-
-__all__ = [
-    'query_tempo_tool',
-    'get_trace_details_tool', 
-    'chat_tempo_tool',
-    'TempoQueryTool',
-    'TempoQueryService',
-    'TempoTraceFormatter',
-    'TempoTraceAnalyzer'
-]
